[Refactor] Move PipelineBuilderContext into ExecRuntime
