+++ be/src/exec/intersect_node.cpp
@@ -20,6 +20,7 @@
 #include "exec/pipeline/exec_node_pipeline_adapter.h"
 #include "exec/pipeline/limit_operator.h"
 #include "exec/pipeline/pipeline_builder.h"
+#include "exec/pipeline/pipeline_builder_operators.h"
 #include "exec/pipeline/set/intersect_build_sink_operator.h"
 #include "exec/pipeline/set/intersect_context.h"
 #include "exec/pipeline/set/intersect_output_source_operator.h"
@@ -95,11 +96,11 @@ StatusOr<pipeline::OpFactories> IntersectNode::decompose_to_pipeline(pipeline::P
     // Use the first child to build the hast table by IntersectBuildSinkOperator.
     ASSIGN_OR_RETURN(auto ops_with_intersect_build_sink, child(0)->decompose_to_pipeline(context));
     if (_local_partition_by_exprs.empty()) {
-        ops_with_intersect_build_sink = context->maybe_interpolate_local_shuffle_exchange(
-                runtime_state(), id(), ops_with_intersect_build_sink, _child_expr_lists[0]);
+        ops_with_intersect_build_sink = ::starrocks::pipeline::builder::maybe_interpolate_local_shuffle_exchange(
+                context, runtime_state(), id(), ops_with_intersect_build_sink, _child_expr_lists[0]);
     } else {
-        ops_with_intersect_build_sink = context->maybe_interpolate_local_bucket_shuffle_exchange(
-                runtime_state(), id(), ops_with_intersect_build_sink, _local_partition_by_exprs[0]);
+        ops_with_intersect_build_sink = ::starrocks::pipeline::builder::maybe_interpolate_local_bucket_shuffle_exchange(
+                context, runtime_state(), id(), ops_with_intersect_build_sink, _local_partition_by_exprs[0]);
     }
     ops_with_intersect_build_sink.emplace_back(std::make_shared<IntersectBuildSinkOperatorFactory>(
             context->next_operator_id(), id(), intersect_partition_ctx_factory, _child_expr_lists[0],
@@ -113,11 +114,13 @@ StatusOr<pipeline::OpFactories> IntersectNode::decompose_to_pipeline(pipeline::P
     for (size_t i = 1; i < _children.size(); i++) {
         ASSIGN_OR_RETURN(auto ops_with_intersect_probe_sink, child(i)->decompose_to_pipeline(context));
         if (_local_partition_by_exprs.empty()) {
-            ops_with_intersect_probe_sink = context->maybe_interpolate_local_shuffle_exchange(
-                    runtime_state(), id(), ops_with_intersect_probe_sink, _child_expr_lists[i]);
+            ops_with_intersect_probe_sink = ::starrocks::pipeline::builder::maybe_interpolate_local_shuffle_exchange(
+                    context, runtime_state(), id(), ops_with_intersect_probe_sink, _child_expr_lists[i]);
         } else {
-            ops_with_intersect_probe_sink = context->maybe_interpolate_local_bucket_shuffle_exchange(
-                    runtime_state(), id(), ops_with_intersect_probe_sink, _local_partition_by_exprs[i]);
+            ops_with_intersect_probe_sink =
+                    ::starrocks::pipeline::builder::maybe_interpolate_local_bucket_shuffle_exchange(
+                            context, runtime_state(), id(), ops_with_intersect_probe_sink,
+                            _local_partition_by_exprs[i]);
         }
         ops_with_intersect_probe_sink.emplace_back(std::make_shared<IntersectProbeSinkOperatorFactory>(
                 context->next_operator_id(), id(), intersect_partition_ctx_factory, _child_expr_lists[i], i - 1));
