+++ be/src/exec/olap_meta_scan_node.cpp
@@ -20,6 +20,7 @@
 #include "exec/pipeline/scan/meta_scan_operator.h"
 #include "exec/pipeline/scan/morsel_queue_factory.h"
 #include "exec/pipeline/scan/olap_meta_scan_prepare_operator.h"
+#include "runtime/runtime_state.h"
 
 namespace starrocks {
