+++ be/src/exec/pipeline/pipeline_builder_operators.h
@@ -1,0 +1,122 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <functional>
+#include <string>
+#include <tuple>
+#include <vector>
+
+#include "exec/pipeline/exchange/local_exchange.h"
+#include "exec/pipeline/pipeline_fwd.h"
+#include "exec/pipeline/spill_process_channel.h"
+#include "runtime/runtime_state_fwd.h"
+
+namespace starrocks {
+
+class ExprContext;
+
+namespace pipeline {
+
+class PipelineBuilderContext;
+
+namespace builder {
+
+OpFactories maybe_interpolate_local_broadcast_exchange(PipelineBuilderContext* context, RuntimeState* state,
+                                                       int32_t plan_node_id, OpFactories& pred_operators,
+                                                       int num_receivers);
+
+// Input the output chunks from the drivers of pred operators into ONE driver of the post operators.
+OpFactories maybe_interpolate_local_passthrough_exchange(PipelineBuilderContext* context, RuntimeState* state,
+                                                         int32_t plan_node_id, OpFactories& pred_operators);
+OpFactories maybe_interpolate_local_passthrough_exchange(PipelineBuilderContext* context, RuntimeState* state,
+                                                         int32_t plan_node_id, OpFactories& pred_operators,
+                                                         int num_receivers, bool force = false);
+OpFactories maybe_interpolate_local_passthrough_exchange(PipelineBuilderContext* context, RuntimeState* state,
+                                                         int32_t plan_node_id, OpFactories& pred_operators,
+                                                         int num_receivers,
+                                                         LocalExchanger::PassThroughType pass_through_type);
+OpFactories maybe_interpolate_local_random_passthrough_exchange(PipelineBuilderContext* context, RuntimeState* state,
+                                                                int32_t plan_node_id, OpFactories& pred_operators,
+                                                                int num_receivers, bool force = false);
+OpFactories maybe_interpolate_local_adpative_passthrough_exchange(PipelineBuilderContext* context, RuntimeState* state,
+                                                                  int32_t plan_node_id, OpFactories& pred_operators,
+                                                                  int num_receivers, bool force = false);
+
+// using KeyPartitionExchanger
+// interpolate local shuffle exchange with partition_exprs
+OpFactories interpolate_local_key_partition_exchange(PipelineBuilderContext* context, RuntimeState* state,
+                                                     int32_t plan_node_id, OpFactories& pred_operators,
+                                                     const std::vector<ExprContext*>& partition_expr_ctxs,
+                                                     int num_receivers,
+                                                     const std::vector<std::string>& transform_exprs = {});
+
+/// Local shuffle the output chunks from multiple drivers of pred operators into DOP partitions of the post operators.
+/// The partition is generated by evaluated each row via partition_expr_ctxs.
+/// When interpolating a local shuffle?
+/// - Local shuffle is interpolated only when DOP > 1 and the source operator of pred_operators could local shuffle.
+/// partition_exprs
+/// - If the source operator has a partition exprs, use it as partition_exprs.
+/// - Otherwise, use self_partition_exprs or self_partition_exprs_generator().
+OpFactories maybe_interpolate_local_shuffle_exchange(PipelineBuilderContext* context, RuntimeState* state,
+                                                     int32_t plan_node_id, OpFactories& pred_operators,
+                                                     const std::vector<ExprContext*>& self_partition_exprs);
+using PartitionExprsGenerator = std::function<std::vector<ExprContext*>()>;
+OpFactories maybe_interpolate_local_shuffle_exchange(PipelineBuilderContext* context, RuntimeState* state,
+                                                     int32_t plan_node_id, OpFactories& pred_operators,
+                                                     const PartitionExprsGenerator& self_partition_exprs_generator);
+
+OpFactories maybe_interpolate_local_bucket_shuffle_exchange(PipelineBuilderContext* context, RuntimeState* state,
+                                                            int32_t plan_node_id, OpFactories& pred_operators,
+                                                            const std::vector<ExprContext*>& partition_expr_ctxs);
+// The intput data is already ordered by partition_exprs. Then we can use a simply approach to split them into different
+// channels as long as the data of the same partition_exprs are in the same channel.
+OpFactories maybe_interpolate_local_ordered_partition_exchange(PipelineBuilderContext* context, RuntimeState* state,
+                                                               int32_t plan_node_id, OpFactories& pred_operators,
+                                                               const std::vector<ExprContext*>& partition_expr_ctxs);
+
+void interpolate_spill_process(PipelineBuilderContext* context, size_t plan_node_id,
+                               const SpillProcessChannelFactoryPtr& channel_factory, size_t dop);
+
+OpFactories interpolate_grouped_exchange(PipelineBuilderContext* context, int32_t plan_node_id,
+                                         OpFactories& pred_operators);
+OpFactories maybe_interpolate_grouped_exchange(PipelineBuilderContext* context, int32_t plan_node_id,
+                                               OpFactories& pred_operators);
+
+// Uses local exchange to gather the output chunks of multiple predecessor pipelines
+// into a new pipeline, which the successor operator belongs to.
+// Append a LocalExchangeSinkOperator to the tail of each pipeline.
+// Create a new pipeline with a LocalExchangeSourceOperator.
+// These local exchange sink operators and the source operator share a passthrough exchanger.
+// pass_through_type specify how to move chunk from prev pipeline to new pipeline.
+OpFactories maybe_gather_pipelines_to_one(PipelineBuilderContext* context, RuntimeState* state, int32_t plan_node_id,
+                                          std::vector<OpFactories>& pred_operators_list,
+                                          LocalExchanger::PassThroughType pass_through_type);
+
+OpFactories maybe_interpolate_collect_stats(PipelineBuilderContext* context, RuntimeState* state, int32_t plan_node_id,
+                                            OpFactories& pred_operators);
+
+OpFactories maybe_interpolate_debug_ops(PipelineBuilderContext* context, RuntimeState* state, int32_t plan_node_id,
+                                        OpFactories& pred_operators);
+
+bool should_interpolate_cache_operator(PipelineBuilderContext* context, int32_t plan_node_id, OpFactoryPtr& source_op);
+OpFactories interpolate_cache_operator(
+        PipelineBuilderContext* context, int32_t plan_node_id, OpFactories& upstream_pipeline,
+        OpFactories& downstream_pipeline,
+        const std::function<std::tuple<OpFactoryPtr, SourceOperatorFactoryPtr>(bool)>& merge_operators_generator);
+
+} // namespace builder
+} // namespace pipeline
+} // namespace starrocks
